Updated devtool option to a full source-map (#1354)
* Updated devtool option to a fully source-map

After reviewing the devtool options from Webpack the `cheap-source-map` option does not make much sense when wanting to have sourcemaps in production.

Sourcemaps in production are usually used for error reporting tools to unminify code within your error reporting tool. Tools like:

- Track.js
- Sentry
- Rollbar
- Etc.

The majority of Webpack plugins that provide support for these error reporting services default to `source-map` as the primary production devtool option. The Webpack documentation also specifies the `eval-source-map` option as follows:

> The following options are not ideal for development nor production.

https://webpack.js.org/configuration/devtool/#special-cases

On the other hand for the `source-map` option it states:

> These options are typically used in production

https://webpack.js.org/configuration/devtool/#production

A full source map provides `original code` quality which is needed for services like these, to work as efficiently as possible.

https://webpack.js.org/configuration/devtool/#devtool

If there are cases against using a full `source-map` for the devtool option in production I would love to hear them.

* Update Api.js